pyproject.toml
[tool.poetry]
name = "slow-api"
version = "v0.1.06"
description = "banger."
authors = ["derpinskV2 <kderpinski@proton.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-environ = "^0.11.2"
django-ninja = "^1.3.0"
django-ninja-extra = "^0.21.8"
orjson = "^3.10.12"
pre-commit = "^4.0.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}
setuptools = "^75.6.0"
toml = "^0.10.2"
urllib3 = "^2.3.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
wheel = "^0.45.1"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^24.10.0"}
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
